initialised the flow of redux-toolkit pattern
used createSlice() to generate the actions and reducers in 'features/cake/cakeSlice.js'

created the store using configureStore() method, and then attached the reducer, in 'app/store.js' file.

performed mutations on the state, in 'index.js' file.
Due to the immer library built in the @reduxjs/toolkit package in 'package.son' file.